+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### Changed
+- `P::flip` now returns a curried function.
+
 ## [0.7.0] - 2016-03-27
 ### Added
 - PHP 7.1 test coverage